+++ NEWS.md
@@ -1,4 +1,4 @@
-# spant 0.8.0
+# spant 0.9.0
 * Updated plotting modes to be one of : "re", "im", "mod" or "arg".
 * Updated int_spec function to use "re", "im", or "mod".
 * Added a function to replicate data across a particular dimension.